refactor(server): /api/jobs CRUD + retry 라우트 분할 (Plan C #C9 4단계)
- src/server/routes/jobs.ts 신규: 6개 라우트 캡슐화 (list, get, cancel, priority, delete, retry)
- safeError 적용 (sanitizeErrorMessage(getErrorMessage) 치환)
- dashboard-api.ts에서 unused import 정리 (ListJobsOptions, GetJobsQuerySchema,
  CancelJobRequestSchema, UpdateJobPriorityRequestSchema, RetryJobRequestSchema,
  GetNotificationsQuerySchema)
- dashboard-api.ts 1593 → 1480줄 (-113)

비포함: /api/jobs/:id/logs/stream (SSE — sse 도메인 PR에서 처리)
남은 도메인: config/setup/stats/repositories/health/doctor/sse
